RGW - Zipper - Add rgw_sal_fwd.h
Add a header that forward declares the SAL objects.  This can be used in
various headers in RGW.
